added video to video folder
