+++ video/video.py
@@ -1,0 +1,10 @@
+
+class Video():
+    def __init__(self, default_settings) -> None:
+        self.user_settings = default_settings
+        self.video_name = ""
+        self.video_path = ""
+        self.video_date_created = ""
+        self.video_length = ""
+        self.video_directory = ""
+
